Split notification support into a separate file.
